% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpatialIMIXObject.R
\name{CreateSpatialIMIXObject}
\alias{CreateSpatialIMIXObject}
\title{CreateSpatialIMIXObject}
\usage{
CreateSpatialIMIXObject(
  ratio,
  location,
  label,
  reference_label = "LG",
  second_label = "HG",
  top_label = "UC"
)
}
\arguments{
\item{ratio}{An n x d data frame or matrix of the log2ratio between the samples and the controls}

\item{location}{Sample corrdinates to compute the kernel matrix, d x 2 data matrix (d is the number of samples, the columns correspond to the x and y coordinates of the sample on a 2-dimentional space)}

\item{label}{A vector of lenghth d corresponding to the sample subtypes of the samples}

\item{reference_label}{The reference label for input in the spatial mixed model, here the example is LG (low-grade intraurothelial neoplasia/normal-appearing urothelial)}

\item{second_label}{The second layer label for input in the spatial mixed model, here the example is HG (high-grade intraurothelial neoplasia)}

\item{top_label}{The top layer label for input in the spatial mixed model, here the example is UC (urothelial carcinoma)}
}
\value{
The data input ready for spatial IMIX
 \item{IMIX_Input_Zscores}{Z scores input for IMIX model, this is inverse normal transformed from the p-values using LRT test on the spatial mixed model results}
 \item{spatial_mixed_model_result}{A data matrix of results fitting the spatial mixed model, including the estimated values, SE, DF, t-value, p-value for each gene of LG, HG, UC. The output LG, HG, UC correspond to reference_label, second_label, top_label respectively}
 \item{full_model}{List of results fitting the full spatial mixed model with variables including HG, UC and the intercept, these are used to construct the LRT test statistics and p-values}
 \item{model_withouthg}{List of results fitting the spatial mixed model with the intercept and UC but without HG, these are used to construct the LRT test statistics and p-values}
 \item{model_withoutintercept}{List of results fitting the spatial mixed model with HG and UC but without the intercept, these are used to construct the LRT test statistics and p-values}
 \item{model_withoutuc}{List of results fitting the spatial mixed model with the intercept and HG but without UC, these are used to construct the LRT test statistics and p-values}
}
\description{
Create Data Input for IMIX model for one data type
}
\examples{
# A toy example
data("example_spatial_IMIX")
imix_object_datatype1 <- CreateSpatialIMIXObject(ratio=ratio1,label=label,location=location)
imix_object_datatype2 <- CreateSpatialIMIXObject(ratio=ratio2,label=label,location=location)
}
