% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_imix.R
\name{fit_multi}
\alias{fit_multi}
\title{Fit Multivariate IMIX model}
\usage{
fit_multi(input1, input2, maxiter = 1000, seed0 = 10)
}
\arguments{
\item{input1}{This is the output 'IMIX_Input_Zscores' after using 'CreateSpatialIMIXObejct' function on data type 1. An n x d data frame or matrix of the summary statistics z score, n is the nubmer of genes, d is the number of labels. Each row is a gene, each column is a label within data type 1.}

\item{input2}{This is the output 'IMIX_Input_Zscores' after using 'CreateSpatialIMIXObejct' function on data type 2. An n x d data frame or matrix of the summary statistics z score, n is the nubmer of genes, d is the number of labels. Each row is a gene, each column is a label within data type 2.}

\item{maxiter}{The maximum number of iteration, default is 1000}

\item{seed0}{set.seed, default is 10}
}
\value{

}
\description{
This function assigns initial values for the input for multivariate IMIX.
}
\examples{
test_multi=fit_multi(input1=imix_object_datatype1$IMIX_Input_Zscores,input2=imix_object_datatype2$IMIX_Input_Zscores)
}
