% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMIX_spatial.R
\name{imix_spatial}
\alias{imix_spatial}
\title{Fit IMIX Spatial Model}
\usage{
imix_spatial(
  input1,
  input2,
  model_type = c("univariate", "multivariate"),
  input_initial_model,
  g_spatial = 64,
  tol0 = 1e-06,
  maxiter0 = 1000,
  seed0 = 10,
  verbose0 = FALSE
)
}
\arguments{
\item{input1}{This is the output 'IMIX_Input_Zscores' after using 'CreateSpatialIMIXObejct' function on data type 1. An n x d data frame or matrix of the summary statistics z score, n is the nubmer of genes, d is the number of labels. Each row is a gene, each column is a label within data type 1.}

\item{input2}{This is the output 'IMIX_Input_Zscores' after using 'CreateSpatialIMIXObejct' function on data type 2. An n x d data frame or matrix of the summary statistics z score, n is the nubmer of genes, d is the number of labels. Each row is a gene, each column is a label within data type 2.}

\item{seed0}{set.seed, default is 10}

\item{maxiter}{The maximum number of iteration, default is 1000}
}
\value{

}
\description{
This function fits the spatial IMIX model
}
\examples{
fit_imix_spatial=imix_spatial(input1=imix_object_datatype1$IMIX_Input_Zscores,input2=imix_object_datatype2$IMIX_Input_Zscores,model_type = "univariate",input_initial_model = test_uni)
fit_imix_spatial_multi=imix_spatial(input1=imix_object_datatype1$IMIX_Input_Zscores,input2=imix_object_datatype2$IMIX_Input_Zscores,model_type = "multivariate",input_initial_model = test_multi)
}
