% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label}
\alias{label}
\title{Example label information for spatial genomic data for two data types}
\format{
A vector of 34 samples:
}
\usage{
data(label)
}
\description{
An example dataset with pathology subtypes/disease grades for the geographical regions, assume both data types have the same subtype information for the samples
}
\examples{
data(label)
}
\keyword{datasets}
