% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location}
\alias{location}
\title{Example location data for spatial genomic data for two data types}
\format{
A data matrix with 34 samples and 2 varialbes:
\describe{
\item{Column}{The column information for each sample}
\item{Row}{The row information for each sample}
}
}
\usage{
data(location)
}
\description{
An example dataset with location for the geographical regions, assume both data types have the same location information for the samples
}
\examples{
data(location)
}
\keyword{datasets}
