% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ratio2}
\alias{ratio2}
\title{Example ratio data for spatial genomic data for data type 2}
\format{
A data matrix with 100 genes and 34 samples, rows are gene names and columns are the geographical region/sample names
}
\usage{
data(ratio2)
}
\description{
An example dataset with the log2ratio compared to healthy controls for the geographical regions for data type 2
}
\examples{
data(ratio2)
}
\keyword{datasets}
