% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_spatial.R
\name{summary_spatial}
\alias{summary_spatial}
\title{Summarize the results after fitting IMIX Spatial Model}
\usage{
summary_spatial(
  input_spatialmodel,
  threshold = 0.1,
  imix_object_datatype1,
  imix_object_datatype2
)
}
\arguments{
\item{input_spatialmodel}{This is the output 'IMIX_Input_Zscores' after using 'imix_spatial' function}

\item{threshold}{The desired FDR threshold for the results, default is 0.1}
}
\value{

}
\description{
This function maps the results after fitting the spatial IMIX model for the two data types to field effect genes
}
\examples{
res=summary_spatial(fit_imix_spatial,imix_object_datatype1 = imix_object_datatype1,imix_object_datatype2 = imix_object_datatype2)
}
